nixos/services/zerotierone: add setting to auto-configure DNS
This commit adds the boolean
services.zerotierone.generateSystemdNetworkdConfig
(and the related services.zerotierone.systemdManagerPackage)
which enables running zerotier-systemd-manager after starting zerotier.
This is done in postStart, because the connection needs to be established
for the zerotier-systemd-manager to work. Currently, this is done with
a sleep command. There may be a better way, but I think this suffices.